+++ changelog.md
@@ -4,6 +4,7 @@
  * move most function related to IO of the command line wrappers into a separate file.
  * make TreeTime own its random number generator and add `--rgn-seed` to control state in CLI. See [PR #234](https://github.com/neherlab/treetime/pull/234)
  * add flag `--greedy-resolve` as inverse to `--stochastic-resolve` with the aim of switching the two.
+ * tighten conditions that trigger approximation of narrow distribution as a delta function in convolution using FFT [PR #235](https://github.com/neherlab/treetime/pull/235).
 
 # 0.9.6: bug fixes and new mode of polytomy resolution
  * in cases when very large polytomies are resolved, the multiplication of the discretized message results in messages/distributions of length 1. This resulted in an error, since interpolation objects require at least two points. This is now caught and a small discrete grid created.