+++ changelog.md
@@ -2,7 +2,7 @@
 
  * the output directory now contains a json file that is compatible with auspice.us. Both time scaled phylogenies and ancestral inferences can now be visualized and explored using auspice. Available colorings are "Date", "genotype", "Branch support", and Excluded. See [PR #232](https://github.com/neherlab/treetime/pull/232) for details.
  * move most function related to IO of the command line wrappers into a separate file.
- * make TreeTime own its random number generator and add `--rgn-seed` to control state in CLI. See [PR #234](https://github.com/neherlab/treetime/pull/234)
+ * make TreeTime own its random number generator and add `--rng-seed` to control state in CLI. Any previous usage of `numpy.random.seed` will now be ignored in favor of `--rng-seed`. See [PR #234](https://github.com/neherlab/treetime/pull/234)
  * add flag `--greedy-resolve` as inverse to `--stochastic-resolve` with the aim of switching the two.
 
 # 0.9.6: bug fixes and new mode of polytomy resolution